% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import_results.R
\name{format_df_data}
\alias{format_df_data}
\title{format_df_data}
\usage{
format_df_data(df)
}
\arguments{
\item{df}{Input dataframe.}
}
\value{
Updated dataframe.
}
\description{
Formats water quality data for use in app. Must run
  `QAQC_results` first.
}
