% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import_results.R
\name{format_df_score}
\alias{format_df_score}
\title{format_results}
\usage{
format_df_score(df, default_state = NA)
}
\arguments{
\item{df}{Input dataframe.}

\item{default_state}{Which state to use when calculating score, thresholds.}
}
\value{
Updated dataframe.
}
\description{
Formats water quality data for use in app. Must run
  `format_df_data` first.
}
