% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_results.R
\name{format_results}
\alias{format_results}
\title{format_results}
\usage{
format_results(df, default_state = NA)
}
\arguments{
\item{df}{Input dataframe.}
}
\description{
Formats water quality data for use in app. Must run
  `QAQC_results` first.
}
