% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import_sites.R
\name{format_sites}
\alias{format_sites}
\title{format_sites}
\usage{
format_sites(df)
}
\arguments{
\item{df}{Input dataframe.}
}
\value{
Updated dataframe.
}
\description{
Formats site data for use in the app.
}
