% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_select_dropdown.R
\name{select_dropdown}
\alias{select_dropdown}
\title{select_dropdown}
\usage{
select_dropdown(
  id,
  label,
  choices,
  choice_names = NULL,
  multiple = TRUE,
  maxOptions = NULL
)
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
