% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp_river}
\alias{shp_river}
\title{River polylines (shapefile)}
\format{
A dataframe with 2 columns:
 \describe{
   \item{Label}{Label}
   \item{Geometry}{Shapefile geometry}
 }
}
\usage{
shp_river
}
\description{
Polyline shapefile showing local rivers.
}
\keyword{datasets}
