% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp_watershed}
\alias{shp_watershed}
\title{Watershed polygons (shapefile)}
\format{
A dataframe with 2 columns:
 \describe{
   \item{Label}{Label}
   \item{Geometry}{Shapefile geometry}
 }
}
\usage{
shp_watershed
}
\description{
Polygon shapefile showing local watersheds.
}
\keyword{datasets}
