% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_sites.R
\name{format_sites}
\alias{format_sites}
\title{format_sites}
\usage{
format_sites(df)
}
\arguments{
\item{df}{Input dataframe.}
}
\value{
The return value, if any, from executing the function.
}
\description{
Preps site data for use in the app, saves df_sites and
  df_sites_all.
}
