% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_QAQC_results.R
\name{QAQC_results}
\alias{QAQC_results}
\title{QAQC_results}
\usage{
QAQC_results(df, date_format = NULL)
}
\arguments{
\item{df}{Input dataframe.}

\item{date_format}{Date format as string.}
}
\value{
The return value, if any, from executing the function.
}
\description{
Run quick quality control check on water quality data. Runs the
  following checks:
  * Checks column names, renames columns as needed
  * Checks all mandatory columns are present
  * Checks for missing values
}
