% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_QAQC_sites.R
\name{QAQC_sites}
\alias{QAQC_sites}
\title{QAQC_sites}
\usage{
QAQC_sites(df)
}
\arguments{
\item{df}{Input dataframe.}
}
\value{
The return value, if any, from executing the function.
}
\description{
Run quick quality control check on site data. Runs the
  following checks:
  * Checks column names, renames columns as needed
  * Checks all mandatory columns are present (Site_ID, Site_Name, Latitude,
    Longitude)
  * Checks for missing values
  * Checks for duplicate Site_ID, Latitude/Longitude
  * Checks that Latitude, Longitude are numeric
  * If present, checks that entries in State column are correctly spelled
}
